feat: implement strategy pattern for better separation and reuse of code
